Removed warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20136 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: d49a6cc0f6ef96a1ca264f1758ae1fc40d365dd5

+++ server/peer/de.uniol.inf.is.odysseus.peer.distribute.modify.fragmentation.horizontal.hash/src/de/uniol/inf/is/odysseus/peer/distribute/modify/fragmentation/horizontal/hash/physicaloperator/HashFragmentPO.java
@@ -102,7 +102,7 @@ public class HashFragmentPO<T extends IStreamObject<IMetaAttribute>>
 			
 		} else hashCode = Math.abs(object.hashCode());
 		
-		return (int) (hashCode % this.numFragments);
+		return hashCode % this.numFragments;
 		
 	}
